Move refactor visibility helper out of extracted path
The shared visibility helper is now proposed independently in PR #19565.
This refactor still needs its optimized variant for runtime validation-node
derivation, so deleting the helper outright would break the branch.

Move that refactor-specific implementation to runtimeHiddenUtils instead.
This leaves utils/layout/hiddenUtils.ts absent from the branch, allowing the
main implementation from the extracted PR to be merged later without an
add/add conflict while keeping the refactor's runtime path intact.
